Shadow fixes + Aero Window Snap added
Fixes:
- Solution now runs demo, was broken due to some null objects

Added:
Aero Window Snap
- This is a controlled by a new property of
AeroWindowSnapActive,(default to true)
- Downside: there is a notable flicker on the edges if you are using the
Dark theme. This is due to windows drawing it's own border on deactivate
and activate events, so a hack for now is to redraw the window on those
events. Couldn't figure out a way to stop windows from drawing it's own
borders.